flaskmigrate/requirements.txt
flask
flask-sqlalchemy
bcrypt
python-dotenv==1.0.1